# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"FirebaseFirestore.Timestamp": {
				"description": "A Timestamp represents a point in time independent of any time zone or\ncalendar, represented as seconds and fractions of seconds at nanosecond\nresolution in UTC Epoch time. It is encoded using the Proleptic Gregorian\nCalendar which extends the Gregorian calendar backwards to year one. It is\nencoded assuming all minutes are 60 seconds long, i.e. leap seconds are\n\"smeared\" so that no leap second table is needed for interpretation. Range\nis from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59.999999999Z.",
				"properties": {
					"seconds": {
						"type": "number",
						"format": "double",
						"description": "The number of seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z."
					},
					"nanoseconds": {
						"type": "number",
						"format": "double",
						"description": "The non-negative fractions of a second at nanosecond resolution."
					}
				},
				"required": [
					"seconds",
					"nanoseconds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameSave": {
				"properties": {
					"uid": {
						"type": "string"
					},
					"data": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"gameId": {
						"type": "string"
					},
					"saveName": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"createdAt": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					},
					"updatedAt": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					}
				},
				"required": [
					"uid",
					"data",
					"gameId",
					"saveName",
					"userId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Game": {
				"properties": {
					"uid": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"requirement": {
						"type": "string"
					},
					"views": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"deeplink": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"previewImage": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"downloadLink": {
						"type": "string"
					},
					"androidLink": {
						"type": "string"
					},
					"iosLink": {
						"type": "string"
					},
					"createdAt": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"uid",
					"status",
					"requirement",
					"views",
					"name",
					"slug",
					"description",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Score": {
				"properties": {
					"gameId": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"createdAt": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					},
					"updatedAt": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					}
				},
				"required": [
					"gameId",
					"score",
					"userId",
					"username",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Rank": {
				"properties": {
					"gameId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"rank": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"createdAt": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					},
					"updatedAt": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					}
				},
				"required": [
					"gameId",
					"userId",
					"rank",
					"username",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TicketPrice": {
				"properties": {
					"ticket": {
						"$ref": "#/components/schemas/Ticket"
					},
					"currency": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"currency",
					"price",
					"decimals",
					"address",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Ticket": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"prices": {
						"items": {
							"$ref": "#/components/schemas/TicketPrice"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TicketBalance": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"price": {
						"items": {
							"$ref": "#/components/schemas/TicketPrice"
						},
						"type": "array"
					},
					"balance": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"image",
					"price",
					"balance"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "masters-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/games/saves/{saveId}": {
			"delete": {
				"operationId": "DeleteGameSave",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Game"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "saveId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateGameSave",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameSave"
								}
							}
						}
					}
				},
				"tags": [
					"Game"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "saveId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"data": {
										"properties": {},
										"additionalProperties": {},
										"type": "object"
									}
								},
								"required": [
									"data"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetGameSave",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameSave"
								}
							}
						}
					}
				},
				"tags": [
					"Game"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "saveId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/games/saves": {
			"post": {
				"operationId": "CreateGameSave",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameSave"
								}
							}
						}
					}
				},
				"tags": [
					"Game"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"saveName": {
										"type": "string"
									},
									"data": {
										"properties": {},
										"additionalProperties": {},
										"type": "object"
									},
									"gameId": {
										"type": "string"
									}
								},
								"required": [
									"data",
									"gameId"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllGameSaves",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameSave"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Game"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/games": {
			"get": {
				"operationId": "GetGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Game"
								}
							}
						}
					}
				},
				"tags": [
					"Game"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/leaderboard/scores": {
			"post": {
				"operationId": "AddScore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Score"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"clientId": {
										"type": "string"
									},
									"apiKey": {
										"type": "string"
									},
									"score": {
										"type": "number",
										"format": "double"
									},
									"gameId": {
										"type": "string"
									}
								},
								"required": [
									"score",
									"gameId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/leaderboard/scores/{gameId}": {
			"get": {
				"operationId": "GetScoreLeaderboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Score"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/leaderboard/scores/{gameId}/me": {
			"get": {
				"operationId": "GetUserScore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Score"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/leaderboard/rank": {
			"post": {
				"operationId": "UpdateRank",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Rank"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"clientId": {
										"type": "string"
									},
									"apiKey": {
										"type": "string"
									},
									"rank": {
										"type": "number",
										"format": "double"
									},
									"gameId": {
										"type": "string"
									}
								},
								"required": [
									"rank",
									"gameId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/leaderboard/rank/{gameId}": {
			"get": {
				"operationId": "GetRankLeaderboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"rank": {
												"type": "number",
												"format": "double"
											},
											"userId": {
												"type": "string"
											}
										},
										"required": [
											"rank",
											"userId"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/leaderboard/rank/{gameId}/me": {
			"get": {
				"operationId": "GetUserRank",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Rank"
								}
							}
						}
					}
				},
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/tickets/balance/{address}": {
			"get": {
				"operationId": "GetTicketsBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TicketBalance"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tickets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tickets": {
			"get": {
				"operationId": "GetTickets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Ticket"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tickets"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tickets/{id}": {
			"get": {
				"operationId": "GetTicket",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ticket"
								}
							}
						}
					}
				},
				"tags": [
					"Tickets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "chainId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}